+++ src/main/kotlin/com/github/etkachev/nxwebstorm/actionlisteners/ReFetchSchematicsListener.kt
@@ -4,7 +4,6 @@ import com.github.etkachev.nxwebstorm.ui.SchematicsListToolTab
 import com.github.etkachev.nxwebstorm.utils.FindAllSchematics
 import com.intellij.openapi.wm.ToolWindow
 import com.intellij.ui.content.ContentFactory
-import java.awt.event.ActionEvent
 
 class ReFetchSchematicsListener(
   private val toolWindow: ToolWindow,
@@ -17,7 +16,7 @@ class ReFetchSchematicsListener(
   /**
    * gets the action listener re-fetching schematics.
    */
-  fun getActionListener(removeOldListener: () -> Unit): (ActionEvent) -> Unit {
+  fun getActionListener(removeOldListener: () -> Unit): () -> Unit {
     return { fetchSchematics(removeOldListener) }
   }